Put wallet address in a separate action file
